+++ src/modules/categorys/categorys.module.ts
@@ -1,6 +1,6 @@
 import { Module } from '@nestjs/common';
 import { CustomTypeOrmModule } from '../../config/typeorm/custom-typeorm-module';
-import { CategorysRepository } from './categorys.repository';
+import { CategorysRepository } from './categorys-typeorm.repository';
 import { CategorysController } from './categorys.controller';
 import { CategorysService } from './categorys.service';
