+++ src/modules/posts/posts.module.ts
@@ -7,7 +7,7 @@ import { TypeOrmModule } from '@nestjs/typeorm';
 import { User } from '../users/entities/user.entity';
 import { Category } from '../categorys/entities/category.entity';
 import { UsersTypeormRepository } from '../users/users-typeorm.repository';
-import { CategorysRepository } from '../categorys/categorys.repository';
+import { CategorysRepository } from '../categorys/categorys-typeorm.repository';
 import { PostLike } from './entities/post-like.entity';
 import { PostsRepository } from './posts.repository';
 import { PrismaModule } from '../../prisma/prisma.module';
