+++ src/test/repository/categorys.repository.spec.ts
@@ -1,7 +1,7 @@
 import { Test, TestingModule } from '@nestjs/testing';
 import { CustomTypeOrmModule } from '../../config/typeorm/custom-typeorm-module';
 import { Category } from '../../modules/categorys/entities/category.entity';
-import { CategorysRepository } from '../../modules/categorys/categorys.repository';
+import { CategorysRepository } from '../../modules/categorys/categorys-typeorm.repository';
 import { TestTypeOrmModule } from '../db/test-db.module';
 
 describe('CategoryRepository', () => {
