+++ src/test/service/posts.service.spec.ts
@@ -7,7 +7,7 @@ import { CreatePostDto } from '../../modules/posts/dto/create-post.dto';
 import { Category } from '../../modules/categorys/entities/category.entity';
 import { NotFoundException, UnauthorizedException } from '@nestjs/common';
 import { UsersRepository } from '../../modules/users/users-typeorm.repository';
-import { CategorysRepository } from '../../modules/categorys/categorys.repository';
+import { CategorysRepository } from '../../modules/categorys/categorys-typeorm.repository';
 import { Post } from '../../modules/posts/entities/post.entity';
 import { UpdatePostDto } from '../../modules/posts/dto/update-post.dto';
 import { TestTypeOrmModule } from '../db/test-db.module';
